% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_00_filenames}
\alias{calmet_00_filenames}
\title{Set input and output filenames for CALMET model run}
\usage{
calmet_00_filenames(calmet_inp = "calmet_template.txt", geodat = NULL,
  srfdat = NULL, clddat = NULL, prcdat = NULL, wtdat = NULL,
  metlst = NULL, metdat = NULL, pacdat = NULL, lcfiles = TRUE,
  nusta = NULL, nowsta = NULL, nm3d = NULL, nigf = NULL, updat = NULL,
  seadat = NULL, m3ddat = NULL, igfdat = NULL, diadat = NULL,
  prgdat = NULL, tstprt = NULL, tstout = NULL, tstkin = NULL,
  tstfrd = NULL, tstslp = NULL, dcstgd = NULL)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}
}
\description{
This function sets the input and output filenames for CALMET model runs in the working CALMET.INP file.
}

