% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_01_temporal_params}
\alias{calmet_01_temporal_params}
\title{Set CALMET temporal and model run parameters}
\usage{
calmet_01_temporal_params(calmet_inp = "calmet_template.txt",
  read_data_from_surf_dat = TRUE, ibyr = NULL, ibmo = NULL, ibdy = NULL,
  ibhr = NULL, ibsec = 0, ieyr = NULL, iemo = NULL, iedy = NULL,
  iehr = NULL, iesec = 0, abtz = NULL, nsecdt = 3600, irtype = 1,
  lcalgrd = TRUE, itest = 2, mreg = 0)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}

\item{read_data_from_surf_dat}{}

\item{ibyr}{}

\item{ibmo}{}

\item{ibdy}{}

\item{ibhr}{}

\item{ibsec}{}

\item{ieyr}{}

\item{iemo}{}

\item{iedy}{}

\item{iehr}{}

\item{iesec}{}

\item{abtz}{}

\item{nsecdt}{}

\item{irtype}{}

\item{lcalgrd}{}

\item{itest}{}

\item{mreg}{}
}
\description{
This function validates and writes CALMET parameters for the model starting and ending times, the time zone, and model timings.
}

