% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calmet_02_grid_levels.R
\name{calmet_02_grid_levels}
\alias{calmet_02_grid_levels}
\title{Set CALMET parameters for the map projection, datum, and grid definitions}
\usage{
calmet_02_grid_levels(calmet_inp = "calmet_template.txt",
  read_xy_from_geo_dat = TRUE, pmap = "UTM", feast = 0, fnorth = 0,
  iutmzn = NULL, utmhem = NULL, rlat0 = "40N", rlon0 = "90W",
  xlat1 = "30N", xlat2 = "60N", datum = "WGS-84", nx = NULL,
  ny = NULL, dgridkm = NULL, xorigkm = NULL, yorigkm = NULL, nz = 12,
  zface = c(0, 20, 40, 80, 100, 150, 200, 300, 400, 800, 1400, 2000, 3000))
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}

\item{read_xy_from_geo_dat}{}

\item{pmap}{the projection of the CALMET domain.}

\item{feast}{the false easting at the projection origin for TTM, LCC, or LAZA projection types.}

\item{fnorth}{the false northing at the projection origin for TTM, LCC, or LAZA projection types.}

\item{iutmzn}{the UTM zone.}

\item{utmhem}{the UTM hemisphere.}

\item{rlat0}{latitude (decimal degrees) of projection origin for TTM, LCC, PS, EM, or LAZA projection types.}

\item{rlon0}{longitude (decimal degrees) of projection origin for TTM, LCC, PS, EM, or LAZA projection types.}

\item{xlat1}{the lower matching parallel of latitude (decimal degrees) for LCC or PS projection types.}

\item{xlat2}{the upper matching parallel of latitude (decimal degrees) for LCC or PS projection types.}

\item{datum}{the datum-region for output coordinates.}

\item{nx}{the number of grid cells in the x direction.}

\item{ny}{the number of grid cells in the y direction.}

\item{dgridkm}{the grid spacing in units of km.}

\item{xorigkm}{the reference grid x coordinate (in km) of the southwest corner of grid cell (1, 1).}

\item{yorigkm}{the reference grid y coordinate (in km) of the southwest corner of grid cell (1, 1).}

\item{nz}{the number of vertical levels.}

\item{zface}{a vector containing cell face heights in meters.}
}
\description{
This function validates and writes CALMET parameters for the map projection, datum, and grid definitions to the working CALMET.INP file.
}

