% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_06_mixhgt_temp_precip_params}
\alias{calmet_06_mixhgt_temp_precip_params}
\title{Set CALMET parameters for mixing height, temperature, and precipitation}
\usage{
calmet_06_mixhgt_temp_precip_params(calmet_inp = "calmet_template.txt",
  constb = 1.41, conste = 0.15, constn = 2400, constw = 0.16,
  fcoriol = 1e-04, iavezi = 1, mnmdav = 10, hafang = 30, ilevzi = 1,
  imixh = 1, threshl = 0.05, threshw = 0.05, itwprog = 0,
  iluoc3d = 16, dptmin = 0.001, dzzi = 200, zimin = 50, zimax = 3000,
  ziminw = 50, zimaxw = 3000, icoare = 10, dshelf = 0, iwarm = 0,
  icool = 0, itprog = 0, irad = 1, tradkm = 500, numts = 5,
  iavet = 1, tgdefb = -0.0098, tgdefa = -0.0045, jwat1 = 55,
  jwat2 = 55, nflagp = 2, sigmap = 100, cutp = 0.01)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}

\item{constb}{empirical mixing height constant for the neutral, mechanical equation.}

\item{conste}{empirical mixing height constant for the convective mixing height equation.}

\item{constn}{empirical mixing height constant for the stable mixing height equation.}

\item{constw}{empirical mixing height constant for the overwater mixing height equation.}

\item{fcoriol}{absolute value of the Coriolis parameter.}

\item{iavezi}{conduct spatial averaging of mixing heights?}

\item{mnmdav}{maximum search radius (in grid cell units) for averaging of mixing heights.}

\item{hafang}{half-angle of upwind looking cone for averaging of mixing heights.}

\item{ilevzi}{layer of winds used in upwind averaging of mixing heights.}

\item{imixh}{method for computation of the convective mixing height: (1) Maul-Carson for land and water cells, (-1) Maul-Carson for land and water cells and OCD mixing heights overwater, (2) Batchvarova and Gryning for land and water cells, and (-2) Batchvarova and Gryning for land and water cells and OCD mixing heights overwater.}

\item{threshl}{threshold buoyancy flux required to sustain convective mixing height growth overland. Expressed as a heat flux per meter of boundary layer in units of W/m3.}

\item{threshw}{threshold buoyancy flux required to sustain convective mixing height growth overwater. Expressed as a heat flux per meter of boundary layer in units of W/m3.}

\item{itwprog}{option for overwater lapse rates used in convective mixing height growth: (0) use SEA.DAT lapse rates and deltaT (or assume neutral conditions if missing), (1) use prognostic lapse rates (only if IPROG > 2) and SEA.DAT deltaT (or neutral if missing), (2) use prognostic lapse rates and prognostic delta T (only if IPROG > 12 and 3D.DAT version 2.0 or higher).}

\item{iluoc3d}{land use category for ocean in 3D.DAT datasets.}

\item{dptmin}{minimum potential temperature lapse rate in the stable layer above the current convective mixing height (in units of K/m).}

\item{dzzi}{depth of layer (in meters) above current convective mixing height through which lapse rate is computed.}

\item{zimin}{minimum overland mixing height (in meters).}

\item{zimax}{maximum overland mixing height (in meters).}

\item{ziminw}{minimum overwater mixing height (in meters).}

\item{zimaxw}{maximum overwater mixing height (in meters).}

\item{icoare}{method for computation of overwater surface fluxes: (0) original deltaT method, (10) COARE with no wave parameterization (jwave = 0), (11) COARE with wave option (jwave = 1) and default wave properties, (-11) COARE with wave option (jwave = 1) and observed wave properties (must be in SEA.DAT files), (12) COARE with wave option 2 (Taylor and Yelland) and default wave properties, and (-12) COARE with wave option 2 (Taylor and Yelland) and observed wave properties (must be in SEA.DAT files).}

\item{dshelf}{coastal/shallow water length scale (in kilometers) for modified z0 in shallow water.}

\item{iwarm}{COARE warm layer computation: (0) off, (1) on.}

\item{icool}{COARE cool skin layer computation: (0) off, (1) on.}

\item{itprog}{selection of temperature data from observations or prognostic data: (0) use both surface and upper air stations, (1) use both surface stations and MM5/3D files for upper air data (no upper air observations), (2) MM5/3D files for both the surface and upper air data (no surface or upper air observations).}

\item{irad}{interpolation type: (1) 1/R, (2) 1/R^2.}

\item{tradkm}{radius of influence (in kilometers) for temperature interpolation.}

\item{numts}{maximum number of stations to include in temperature interpolation.}

\item{iavet}{conduct spatial averaging of temperatures?}

\item{tgdefb}{default temperature gradient (in units of K/m) below the mixing height over water.}

\item{tgdefa}{default temperature gradient (in units of K/m) above the mixing height over water.}

\item{jwat1}{beginning land use category value for temperature interpolation over water.}

\item{jwat2}{ending land use category value for temperature interpolation over water.}

\item{nflagp}{method of interpolation for precipitation: (1) 1/R, (2) 1/R^2, (3) exp/R^2.}

\item{sigmap}{radius of influence (in kilometers) for interpolation of precipitation.}

\item{cutp}{minimum precipitation rate cutoff (in units of mm/h).}
}
\description{
This function validates and writes CALMET parameters for wind field options and parameters to the working CALMET.INP file.
}

