% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_07_station_params}
\alias{calmet_07_station_params}
\title{Set CALMET surface station, precipitation station, and upper air parameters}
\usage{
calmet_07_station_params(calmet_inp = "calmet_template.txt",
  read_data_from_files = TRUE, surf_station_vars_name = NULL,
  surf_station_vars_ID = NULL, surf_station_vars_x_coord = NULL,
  surf_station_vars_y_coord = NULL, surf_station_vars_tz = NULL,
  surf_station_vars_anem_hgt = NULL, upper_air_params_name = NULL,
  upper_air_params_ID = NULL, upper_air_params_x_coord = NULL,
  upper_air_params_y_coord = NULL, upper_air_params_tz = NULL,
  precip_station_params_name = NULL, precip_station_params_stn_code = NULL,
  precip_station_params_x_coord = NULL,
  precip_station_params_y_coord = NULL)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}

\item{read_data_from_files}{}

\item{surf_station_vars_name}{}

\item{surf_station_vars_ID}{}

\item{surf_station_vars_x_coord}{}

\item{surf_station_vars_y_coord}{}

\item{surf_station_vars_tz}{}

\item{surf_station_vars_anem_hgt}{}

\item{upper_air_params_name}{}

\item{upper_air_params_ID}{}

\item{upper_air_params_x_coord}{}

\item{upper_air_params_y_coord}{}

\item{upper_air_params_tz}{}

\item{precip_station_params_name}{}

\item{precip_station_params_stn_code}{}

\item{precip_station_params_x_coord}{}

\item{precip_station_params_y_coord}{}
}
\description{
This function validates and writes CALMET parameters for surface stations, precipitation stations, and upper air soundings to the working CALMET.INP file.
}

