% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_3_output_opts}
\alias{calmet_3_output_opts}
\title{Set CALMET parameters for the output options}
\usage{
calmet_3_output_opts(calmet_inp, lsave = TRUE, iformo = 0, lprint = FALSE,
  iprinf = 1, iuvout = c(1, rep(0, (nz - 1))), iwout = c(1, rep(0, (nz -
  1))), itout = c(1, rep(0, (nz - 1))), stability = FALSE, ustar = FALSE,
  monin = FALSE, mixht = FALSE, wstar = FALSE, precip = FALSE,
  sensheat = FALSE, convzi = FALSE, ldb = FALSE, nn1 = 1, nn2 = 1,
  ldbcst = FALSE, ioutd = FALSE, nzprn2 = TRUE, ipr0 = FALSE,
  ipr1 = FALSE, ipr2 = FALSE, ipr3 = FALSE, ipr4 = FALSE,
  ipr5 = FALSE, ipr6 = FALSE, ipr7 = FALSE, ipr8 = FALSE)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}

\item{lsave}{save meteorological fields in an unformatted output file.}

\item{iformo}{type of unformatted output file: (1) CALPUFF/CALGRID type file [CALMET.DAT], (2) MESOPUFF-II type file [PACOUT.DAT].}

\item{lprint}{whether to print meteorological fields in the iuvout, iwout, and itout options.}

\item{iprinf}{print interval in hours.}

\item{iuvout}{specification of which layers of U,V wind component to print; vector must be the same length as the number of levels (nz).}

\item{iwout}{specification of which layers of the W wind component to print; vector must be the same length as the number of levels (nz).}

\item{itout}{specification of which levels of the 3D temperature field to print; vector must be the same length as the number of levels (nz).}

\item{stability}{whether to print the PGT stability class (will only print if lprint is TRUE).}

\item{ustar}{whether to print the friction velocity (will only print if lprint is TRUE).}

\item{monin}{whether to print the Monin-Obukhov length (will only print if lprint is TRUE).}

\item{mixht}{whether to print the mixing height (will only print if lprint is TRUE).}

\item{wstar}{whether to print the convective velocity scale (will only print if lprint is TRUE).}

\item{precip}{whether to print the precipitation rate (will only print if lprint is TRUE).}

\item{sensheat}{whether to print the sensible heat flux (will only print if lprint is TRUE).}

\item{convzi}{whether to print the convective mixing height (will only print if lprint is TRUE).}

\item{ldb}{whether to print input meteorological data and internal variables.}

\item{nn1}{if ldb is TRUE, the first time step for which debug data are printed.}

\item{nn2}{if ldb is TRUE, the last time step for which debug data are printed.}

\item{ldbcst}{whether to print distance to land internal variables (output will be a .GRD file DCST.GRD).}

\item{ioutd}{whether to print the test/debug wind fields to disk files.}

\item{nzprn2}{number of levels, starting at the surface, to print.}

\item{ipr0}{print the interpolated wind components?}

\item{ipr1}{print the terrain-adjusted surface wind components?}

\item{ipr2}{print the smoothed wind components and the initial divergence fields?}

\item{ipr3}{print the final wind speed and direction fields?}

\item{ipr4}{print the final divergence fields?}

\item{ipr5}{print the winds after kinematic effects are added?}

\item{ipr6}{print the winds after the Froude number adjustment is made?}

\item{ipr7}{print the winds after slope flows are added?}

\item{ipr8}{print the final wind field components?}
}
\description{
This function validates and writes CALMET parameters for the output options to the working CALMET.INP file.
}

