% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_4_met_data_opts}
\alias{calmet_4_met_data_opts}
\title{Set CALMET parameters for the meteorological data options}
\usage{
calmet_4_met_data_opts(calmet_inp, noobs = 0, nssta, npsta, icloud = 0,
  iforms = 2, iformp = 2, iformc = 2)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}
}
\description{
This function validates and writes CALMET parameters for the meteorological data options to the working CALMET.INP file.
}

