% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_6_mixhgt_temp_precip_params}
\alias{calmet_6_mixhgt_temp_precip_params}
\title{Set CALMET parameters for mixing height, temperature, and precipitation}
\usage{
calmet_6_mixhgt_temp_precip_params(calmet_inp, constb, conste, constn, constw,
  fcoriol, iavezi, mnmdav, hafang, ilevzi, imixh, threshl, threshw, itwprog,
  iluoc3d, dptmin, dzzi, zimin, zimax, ziminw, zimaxw, icoare, dshelf, iwarm,
  icool, itprog, irad, tradkm, numts, iavet, tgdefb, tgdefa, jwat1, jwat2,
  nflagp, sigmap, cutp)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}
}
\description{
This function validates and writes CALMET parameters for wind field options and parameters to the working CALMET.INP file.
}

