% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_define_geophys}
\alias{calmet_define_geophys}
\title{Define the CALMET domain and generate a geophysical input file}
\usage{
calmet_define_geophys(lat_dec_deg = NULL, lon_dec_deg = NULL,
  lat_lon_grid_loc = 1, domain_width_m = NULL, domain_height_m = NULL)
}
\arguments{
\item{lat_dec_deg}{the latitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_lon_grid_loc argument.}

\item{lon_dec_deg}{the longitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_lon_grid_loc argument.}

\item{lat_lon_grid_loc}{the location of the lat/long inputs in relation to the domain. Choices are: 1 (center), 2 (lower left), 3 (lower right), 4 (upper left), 5 (upper right).}

\item{domain_width_m}{the desired width of the meteorological domain in meters.}

\item{domain_height_m}{the desired height of the meteorological domain in meters.}
}
\description{
Define the CALMET domain and determine the best gridded values for land use, terrain heights, and micrometeorological parameters for creation of a geophysical input file.
}
\examples{
\dontrun{
# Create a CALMET domain of 100 by 100 km in the Los Angeles area.
# Chosen lat/lon coordinates are for the center of the domain.
calmet_define_domain(lat_dec_deg = 34.050184,
                     lon_dec_deg = -118.253959,
                     lat_lon_grid_loc = 1,
                     domain_width_m = 8000,
                     domain_height_m = 8000)
}
}

