% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calmet_exec.R
\name{calmet_exec}
\alias{calmet_exec}
\title{Run the CALMET model}
\usage{
calmet_exec(calmet_exec)
}
\arguments{
\item{calmet_exec}{if the CALMET executable is in the system path, provide the name of the executable only; otherwise, provide the full path and name of the CALMET executable.}
}
\description{
Run the CALMET model.
}

