% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_get_ncdc_history}
\alias{calmet_get_ncdc_history}
\title{Retrieve the NCDC history data file}
\usage{
calmet_get_ncdc_history(replace.file = FALSE)
}
\arguments{
\item{replace.file}{selecting 'yes' will overwrite history file if it exists in the working directory.}
}
\description{
This function initiates a download of the NCDC surface station history file.
}
\examples{
\dontrun{
# Obtain the NCDC history file
calmet_get_ncdc_history()
}
}

