% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_get_ncdc_station_data}
\alias{calmet_get_ncdc_station_data}
\title{Obtain NCDC station data}
\usage{
calmet_get_ncdc_station_data(start_year = NULL, end_year = NULL,
  bbox_lat_lon = NULL)
}
\arguments{
\item{start_year}{the year for the start of data retrieval.}

\item{end_year}{the year for the end of data retrieval.}

\item{bbox_lat_lon}{a spatial bounding box in projected in lat/lon coordinates}
}
\description{
Obtain NCDC station data for all stations residing in the CALMET domain during a specied time period.
}
\examples{
\dontrun{
# Get 2010 station data for a previously defined domain
calmet_get_ncdc_station_data(start_year = 2010,
                             end_year = 2010,
                             bbox_lat_lon = bbox)
}
}

