% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calmet_seasonal_micrometeorology.R
\name{calmet_seasonal_micrometeorology}
\alias{calmet_seasonal_micrometeorology}
\title{Return a data frame with seasonal micrometeorological parameters for application to gridded land use categories}
\usage{
calmet_seasonal_micrometeorology()
}
\description{
This function produces a data frame containing values for six micrometeorological parameters for eight CALMET land use categories for four seasons.
}

