% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_surface_met}
\alias{calmet_surface_met}
\title{Obtain surface meteorology and generate input files}
\usage{
calmet_surface_met(location_name, year, lat_dec_deg = NULL,
  lon_dec_deg = NULL, lat_lon_grid_loc = 1, domain_width_m = NULL,
  domain_height_m = NULL, cell_resolution_m = 250, time_offset,
  local_archive_dir = NULL, use_CSV_files = NULL)
}
\arguments{
\item{year}{the year for the SURF.DAT file.}

\item{lat_dec_deg}{the latitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_long_grid_loc argument.}

\item{lon_dec_deg}{the longitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_long_grid_loc argument.}

\item{lat_lon_grid_loc}{the location of the lat/long inputs in relation to the domain. Choices are: 1 (center), 2 (lower left), 3 (lower right), 4 (upper left), 5 (upper right).}

\item{domain_width_m}{the width of the meteorological domain in meters.}

\item{domain_height_m}{the height of the meteorological domain in meters.}

\item{cell_resolution_m}{the grid cell resolution in meters.}

\item{time_offset}{the offset from UTC-00:00 in hours.}

\item{local_archive_dir}{a local path containing an archive of gzipped NCDC station data files.}

\item{use_CSV_files}{option to use local, pre-processed CSV files.}
}
\description{
Obtain surface meteorology and generate input files
}
\examples{
\dontrun{
# Generate a surface station file
calmet_surface_met(location_name = 'test',
                   year = 2005,
                   lat_dec_deg = 49.196116,
                   lon_dec_deg = -122.505866,
                   lat_lon_grid_loc = 1,
                   domain_width_m = 117000,
                   domain_height_m = 43250,
                   cell_resolution_m = 500,
                   time_offset = -8)
}
}

