% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_upper_air}
\alias{calmet_upper_air}
\title{Create an upper air data file for CALMET}
\usage{
calmet_upper_air(location_name, year, lat_dec_deg = NULL,
  lon_dec_deg = NULL, lat_lon_grid_loc = 1, domain_width_m = NULL,
  domain_height_m = NULL, cell_resolution_m = 250, time_offset)
}
\arguments{
\item{location_name}{an assigned name for the CALMET domain.}

\item{year}{the year for the UP.DAT file.}

\item{lat_dec_deg}{the latitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_long_grid_loc argument.}

\item{lon_dec_deg}{the longitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_long_grid_loc argument.}

\item{lat_lon_grid_loc}{the location of the lat/long inputs in relation to the domain. Choices are: 1 (center), 2 (lower left), 3 (lower right), 4 (upper left), 5 (upper right).}

\item{domain_width_m}{the width of the meteorological domain in meters.}

\item{domain_height_m}{the height of the meteorological domain in meters.}

\item{cell_resolution_m}{the grid cell resolution in meters.}

\item{time_offset}{the offset from UTC-00:00 in hours.}
}
\description{
This function retrieves upper air sounding data and processes it to generate an upper air input file for CALMET.
}

