% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calpost_get_concentrations_from_time_series_file.R
\name{calpost_get_concentrations_from_time_series_file}
\alias{calpost_get_concentrations_from_time_series_file}
\title{Create a data frame of discrete receptor concentrations}
\usage{
calpost_get_concentrations_from_time_series_file(time_series_file = NULL,
  location_name, source_id, pollutant_id, create_hourly_CSV = TRUE,
  create_hourly_rda = TRUE, return_large_df = FALSE,
  resume_from_set_hour = NULL, autoresume_processing = TRUE,
  autoresume_year = NULL)
}
\arguments{
\item{time_series_file}{a path to a binary time series data file that was generated by CALPOST with time series options set.}

\item{location_name}{the name of the location in which the receptors reside.}

\item{source_id}{the ID value for the source emissions.}

\item{pollutant_id}{the ID value for the emitted pollutant.}

\item{create_hourly_CSV}{an option to create hourly CSV files describing pollutant concentrations at every receptor.}

\item{create_hourly_rda}{an option to create hourly R data (.rda) files describing pollutant concentrations at every receptor.}

\item{return_large_df}{an option to invisibly return a large data frame object.}

\item{resume_from_set_hour}{an option to resume processing of concentrations from a set hour of the year.}

\item{autoresume_processing}{}

\item{autoresume_year}{}
}
\description{
Create a data frame of discrete receptor concentrations using a CALPOST time series outfile file.
}

