% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_01_run_control_params}
\alias{calpuff_01_run_control_params}
\title{Set CALPUFF run control parameters}
\usage{
calpuff_01_run_control_params(calpuff_inp = "calpuff_template.txt",
  run_all_calmet_periods = TRUE, ibyr = NULL, ibmo = NULL, ibdy = NULL,
  ibhr = NULL, ibmin = 0, ibsec = 0, ieyr = NULL, iemo = NULL,
  iedy = NULL, iehr = NULL, iemin = 0, iesec = 0, abtz = NULL,
  nsecdt = 3600, nspec = NULL, nse = NULL, itest = 2, mrestart = 0,
  nrespd = 0, metfm = 1, mprffm = 1, avet = 60, pgtime = 60,
  ioutu = 1, iovers = 2)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{metrun}{the choice of whether to run all periods available in meteorological input file.}

\item{ibyr}{the starting year for the CALPUFF run.}

\item{ibmo}{the starting month for the CALPUFF run.}

\item{ibdy}{the starting day for the CALPUFF run.}

\item{ibhr}{the starting hour for the CALPUFF run.}

\item{ibmin}{the starting minute for the CALPUFF run.}

\item{ibsec}{the starting second for the CALPUFF run.}

\item{ieyr}{the ending year for the CALPUFF run.}

\item{iemo}{the ending month for the CALPUFF run.}

\item{iedy}{the ending day for the CALPUFF run.}

\item{iehr}{the ending hour for the CALPUFF run.}

\item{iemin}{the starting minute for the CALPUFF run.}

\item{iesec}{the ending second for the CALPUFF run.}

\item{abtz}{the time zone for the CALPUFF domain.}

\item{nsecdt}{the number of seconds between time steps.}

\item{nse}{number of stationary emission sources.}

\item{itest}{a flag to stop the model run after the setup phase, which is useful for checking for correctness in model inputs and associated files.}

\item{mrestart}{a flag to interact with restart files.}

\item{nrespd}{the number of periods in the restart output cycle.}

\item{metfm}{the meteorological data format to be used: (1) CALMET binary file (CALMET.DAT), (2) ISC ASCII file (ISCMET.DAT), (3) AUSPLUME ASCII file (PLMMET.MET), (4) CTDM plus tower file (PROFILE.DAT) and surface parameters file (SURFACE.DAT), or (5) AERMET tower file (PROFILE.DAT) and surface parameters file (SURFACE.DAT).}

\item{mprffm}{the meteorological profile data format (ignored unless metfm is either 1, 2, or 3): (1) CTDM plus tower file, or (2) AERMET tower file.}

\item{avet}{the averaging time in minutes.}

\item{pgtime}{the PG averaging time in minutes.}

\item{ioutu}{the output units for binary concentration and flux files: (1) mass units (e.g., g/m3 or g/m2/s), (2) odour units, or (3) radiation units (e.g., Bq/m3 or Bq/m2/s).}

\item{iovers}{the output dataset format version for binary concentration and flux files: (1) version 2.1, or (2) version 2.2.}
}
\description{
This function validates and writes CALPUFF run control parameters.
}

