% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_01_temporal_params}
\alias{calpuff_01_temporal_params}
\title{Set CALPUFF temporal and model run parameters}
\usage{
calpuff_01_temporal_params(calpuff_inp, ibyr, ibmo, ibdy, ibhr, ibsec, ieyr,
  iemo, iedy, iehr, iesec, abtz, nsecdt, irtype, lcalgrd, itest, mreg)
}
\arguments{
\item{calpuff_inp}{}

\item{ibyr}{}

\item{ibmo}{}

\item{ibdy}{}

\item{ibhr}{}

\item{ibsec}{}

\item{ieyr}{}

\item{iemo}{}

\item{iedy}{}

\item{iehr}{}

\item{iesec}{}

\item{abtz}{}

\item{nsecdt}{}

\item{irtype}{}

\item{lcalgrd}{}

\item{itest}{}

\item{mreg}{}
}
\description{
This function validates and writes CALPUFF parameters for the model starting and ending times, the time zone, and model timings.
}

