% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_03_species_list}
\alias{calpuff_03_species_list}
\title{Set the CALPUFF species list}
\usage{
calpuff_03_species_list(calpuff_inp = "calpuff_template.txt",
  species_names_cspec = NULL, species_modelled = NULL,
  species_emitted = NULL, species_ddep = NULL,
  species_group_number = NULL)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{species_names_cspec}{}

\item{species_modelled}{}

\item{species_emitted}{}

\item{species_ddep}{}

\item{species_group_number}{}
}
\description{
This function validates and writes CALPUFF parameters for the model's species list.
}

