% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_06_complex_terrain_inputs}
\alias{calpuff_06_complex_terrain_inputs}
\title{Set the CALPUFF subgrid scale complex terrain inputs}
\usage{
calpuff_06_complex_terrain_inputs(calpuff_inp, nhill, nctrec, mhill, xhill2m,
  zhill2m, xctdmkm, yctdmkm)
}
\arguments{
\item{calpuff_inp}{}

\item{nhill}{}

\item{nctrec}{}

\item{mhill}{}

\item{xhill2m}{}

\item{zhill2m}{}

\item{xctdmkm}{}

\item{yctdmkm}{}
}
\description{
This function validates and writes CALPUFF subgrid scale complex terrain inputs.
}

