% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_07_chem_ddep_gases}
\alias{calpuff_07_chem_ddep_gases}
\title{Set the CALPUFF chemical parameters for dry deposition of gases}
\usage{
calpuff_07_chem_ddep_gases(calpuff_inp = "calpuff_template.txt",
  species_diffusivity = NULL, species_alpha_star = NULL,
  species_reactivity = NULL, species_mesophyll_resistance = NULL,
  species_HL_coefficient = NULL)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{species_diffusivity}{}

\item{species_alpha_star}{}

\item{species_reactivity}{}

\item{species_mesophyll_resistance}{}

\item{species_HL_coefficient}{}
}
\description{
This function validates and writes CALPUFF chemical parameters for dry deposition of gases.
}

