% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_15_line_source_params}
\alias{calpuff_15_line_source_params}
\title{Set one or more CALPUFF line source parameters}
\usage{
calpuff_15_line_source_params(calpuff_inp, nln2 = NULL, nlines = NULL,
  ilnu = NULL, nsln1 = NULL, mxnseg = NULL, nlrise = NULL, xl = NULL,
  hbl = NULL, wbl = NULL, wml = NULL, dxl = NULL, fprimel = NULL,
  lnsource_params = NULL)
}
\arguments{
\item{calpuff_inp}{}

\item{nln2}{}

\item{nlines}{}

\item{ilnu}{}

\item{nsln1}{}

\item{mxnseg}{}

\item{nlrise}{}

\item{xl}{}

\item{hbl}{}

\item{wbl}{}

\item{wml}{}

\item{dxl}{}

\item{fprimel}{}

\item{lnsource_params}{}
}
\description{
This function validates and writes CALPUFF line source parameters.
}

