% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calpuff_16_volume_source_params.R
\name{calpuff_16_volume_source_params}
\alias{calpuff_16_volume_source_params}
\title{Set one or more CALPUFF volume source parameters}
\usage{
calpuff_16_volume_source_params(calpuff_inp = "calpuff_template.txt")
}
\description{
This function validates and writes CALPUFF volume source parameters.
}

