% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_2_technical_opts}
\alias{calpuff_2_technical_opts}
\title{Set CALPUFF technical options}
\usage{
calpuff_2_technical_opts(calpuff_inp, mgauss, mctadj, mctsg, mslug, mtrans,
  mtip, mrise, mbdw, mshear, msplit, mchem, maqchem, mlwc, mwet, mdry, mtilt,
  mdisp, mturbvw, mdisp2, mtauly, mtauadv, mcturb, mrough, mpartl, mpartlba,
  mtinv, mpdf, msgtibl, mbcon, msource, mfog, mreg)
}
\arguments{
\item{calpuff_inp}{}

\item{mgauss}{}

\item{mctadj}{}

\item{mctsg}{}

\item{mslug}{}

\item{mtrans}{}

\item{mtip}{}

\item{mrise}{}

\item{mbdw}{}

\item{mshear}{}

\item{msplit}{}

\item{mchem}{}

\item{maqchem}{}

\item{mlwc}{}

\item{mwet}{}

\item{mdry}{}

\item{mtilt}{}

\item{mdisp}{}

\item{mturbvw}{}

\item{mdisp2}{}

\item{mtauly}{}

\item{mtauadv}{}

\item{mcturb}{}

\item{mrough}{}

\item{mpartl}{}

\item{mpartlba}{}

\item{mtinv}{}

\item{mpdf}{}

\item{msgtibl}{}

\item{mbcon}{}

\item{msource}{}

\item{mfog}{}

\item{mreg}{}
}
\description{
This function validates and writes CALPUFF parameters for the model's technical options.
}

