% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_9_misc_ddep}
\alias{calpuff_9_misc_ddep}
\title{Set miscellaneous CALPUFF parameters for dry deposition}
\usage{
calpuff_9_misc_ddep(calpuff_inp, rcutr, rgr, reactr, nint, iveg)
}
\arguments{
\item{calpuff_inp}{}

\item{rcutr}{}

\item{rgr}{}

\item{reactr}{}

\item{nint}{}

\item{iveg}{}
}
\description{
This function validates and writes miscellaneous CALPUFF parameters for dry deposition.
}

