% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calpuff_add_volume_sources.R
\name{calpuff_add_volume_sources}
\alias{calpuff_add_volume_sources}
\title{Add volume sources to a list for later use in CALPUFF}
\usage{
calpuff_add_volume_sources(src_name, species_name, lat_dec_deg = NULL,
  lon_dec_deg = NULL, x_coord_km = NULL, y_coord_km = NULL,
  UTM_zone = NULL, UTM_hemisphere = NULL, effective_height, base_elev,
  init_sigma_y, init_sigma_z, emission_rate, emission_units)
}
\arguments{
\item{src_name}{the name of the source emitting the species.}

\item{species_name}{the name of the species undergoing emissions.}

\item{lat_dec_deg}{the latitude of the volume source in decimal degrees.}

\item{lon_dec_deg}{the longitude of the volume source in decimal degrees.}

\item{x_coord_km}{the UTM easting value of the volume source in km units.}

\item{y_coord_km}{the UTM northing value of the volume source in km units.}

\item{UTM_zone}{the UTM zone for the volume source.}

\item{UTM_hemisphere}{the UTM hemisphere for the volume source.}

\item{effective_height}{the effective height of the volume source in meters above ground level (m AGL).}

\item{base_elev}{the ground elevation at the location of the volume source in meters above sea level (m ASL).}

\item{init_sigma_y}{the initial sigma y value for the volume source in meters.}

\item{init_sigma_z}{the initial sigma z value for the volume source in meters.}

\item{emission_rate}{the rate of constant emissions from the volume source; units are defined in the 'emission_units' argument.}

\item{emission_units}{the units applied to the value defined in the 'emission_rate' argument. The possible selections are: (1) "g/s", (2) "kg/hr", (3) "lb/hr", (4) "tons/yr", (5) "Odour Unit * m3/s", (6) "Odour Unit * m3/min", (7) "metric tons/yr", (8) "Bq/s", and (9) "GBq/yr".}

\item{UTM_zone}{the UTM zone for the volume source.}

\item{UTM_hemisphere}{the UTM hemisphere for the volume source.}
}
\description{
Add volume sources to a list for later use in CALPUFF
}

