% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_def_recep_from_geophys}
\alias{calpuff_def_recep_from_geophys}
\title{Generate a grid of receptors from a CALMET geophysical input file}
\usage{
calpuff_def_recep_from_geophys(geophys_file = NULL,
  resolution_scale_factor = 0.5, height_above_ground = 2.5,
  download_SRTM = TRUE, SRTM_file_path = NULL)
}
\arguments{
\item{geophys_file}{a path to a geophysical data file generated using the 'calmet_define_geophys" function.}

\item{resolution_scale_factor}{a factor by which the resolution of the receptors relates to the resolution of the CALMET domain.}

\item{height_above_ground}{the chosen height above ground for all the discrete receptors.}

\item{download_SRTM}{a choice of whether to download the SRTM GeoTIFF height data from a server or read the identical files from a local folder.}

\item{SRTM_file_path}{path to a folder containing a collection of SRTM V4 zip archive files.}
}
\description{
Generate a grid of receptors from a pre-existing CALMET geophysical input file.
}

