% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_exec}
\alias{calpuff_exec}
\title{Run the CALPUFF model}
\usage{
calpuff_exec(calpuff_exec)
}
\arguments{
\item{calpuff_exec}{if the CALPUFF executable is in the system path, provide the name of the executable only; otherwise, provide the full path and name of the CALPUFF executable.}
}
\description{
Run the CALPUFF model.
}

