% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download_FSL_sounding_data}
\alias{download_FSL_sounding_data}
\title{Download an upper air data file in the FSL format}
\usage{
download_FSL_sounding_data(sounding_priority, df_soundings,
  station_list_position, starting_hour, level_type, wind_units, beginning_date,
  ending_date)
}
\arguments{
\item{sounding_priority}{a character string that determines whether the sounding is considered as "primary" or "secondary".}

\item{df_soundings}{a data frame containing a complete database of global sounding sites.}

\item{station_list_position}{the row of interest from the 'df_soundings' data frame.}

\item{starting_hour}{the starting hour for the sounding data.}

\item{level_type}{the type of levels to obtain from the FSL upper air sounding data.}

\item{wind_units}{the type of wind units to obtain from the FSL upper air sounding data.}

\item{beginning_date}{the beginning date of the requested sounding data.}

\item{ending_date}{the ending date of the requested sounding data.}
}
\description{
Download an upper air data file from NOAA in the FSL format.
}

