% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download_SRTMV4_GeoTIFF.R
\name{download_SRTMV4_GeoTIFF}
\alias{download_SRTMV4_GeoTIFF}
\title{Download and rasterize SRTM V4 GeoTIFF files}
\usage{
download_SRTMV4_GeoTIFF(SP_object)
}
\arguments{
\item{SP_object}{a SpatialPoints object with a longlat projection.}
}
\description{
Download one or more SRTM V4 GeoTIFF files and merge files as necessary to encompass modelling domain
}

