\name{grid.define.METxy}
\alias{grid.define.METxy}
\title{

}

\description{
Define the meteorological grid surface.
}

\usage{
grid.define.METxy()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
