\name{grid.define.METz}
\alias{grid.define.METz}
\title{

}

\description{
Define the meteorological levels in the z direction.
}

\usage{
grid.define.METz()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
