\name{grid.define.proj}
\alias{grid.define.proj}
\title{

}

\description{
Define grid projection and datum.
}

\usage{
grid.define.proj()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
