\name{grid.define.SAMP}
\alias{grid.define.SAMP}
\title{

}

\description{
Define a sampling grid.
}

\usage{
grid.define.SAMP
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
