\name{grid.summary}
\alias{grid.summary}
\title{

}

\description{
Get a summary of the currently defined grids.
}

\usage{
grid.summary()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
