\name{met_dat_opts.set}
\alias{met_dat_opts.set}
\title{

}

\description{
Set meteorological data options.
}

\usage{
met_dat_opts.set()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
