\name{mixhgt_temp_precip.set}
\alias{mixhgt_temp_precip.set}
\title{

}

\description{
Set mixing height, temperature, and precipitation parameters.
}

\usage{
mixhgt_temp_precip.set()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
