% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ncdc.station.data}
\alias{ncdc.station.data}
\title{Obtain NCDC station data}
\usage{
ncdc.station.data(startyear = NULL, endyear = NULL,
  use.calmet.bounds = TRUE, manual.bounds = NULL)
}
\arguments{
\item{startyear}{the year for the start of data retrieval.}

\item{endyear}{the year for the end of data retrieval.}

\item{use.calmet.bounds}{a 'TRUE' or 'FALSE' option on whether to use the same bounds as the previously defined CALMET domain.}

\item{manual.bounds}{option to specify boundaries using lat/long coordinates in decimal degrees.}
}
\description{
Obtain NCDC station data for all stations residing in the CALMET domain during a specied time period.
}
\examples{
\dontrun{
# Get 2010 station data for a previously defined domain
ncdc.station.data(startyear = 2010,
                  endyear = 2010,
                  use.calmet.bounds = TRUE)
}
}

