\name{plot.calmet.domain}
\alias{plot.calmet.domain}
\title{
Plot all surface meteorological stations in CALMET domain
}

\description{
This function overlays all available surface stations in the NCDC dataset that lie within the defined domain. The temporal element is relevant since only those stations with data for the predefined period with appear in this plot.
}

\usage{
plot.calmet.domain()
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{
# Plot all stations in NCDC database in the defined CALMET domain
plot.calmet.domain()
}
