\name{project.define}
\alias{project.define}
\title{

}

\description{
Define a project, create folder skeleton, init dictionaries.
}

\usage{
project.define()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
