\name{project.list}
\alias{project.list}
\title{

}

\description{
Get a list of all projects in system (either simple or verbose).
}

\usage{
project.list()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
