\name{project.modify}
\alias{project.modify}
\title{

}

\description{
Modify project settings originally set with 'project.define'.
}

\usage{
project.modify()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
