\name{project.summary}
\alias{project.summary}
\title{

}

\description{
Get a summary of the current working project.
}

\usage{
project.summary()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
