\name{project.switch}
\alias{project.switch}
\title{

}

\description{
Switch from the current project to another project.
}

\usage{
project.switch()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
