% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read62_01_run_control_params.R
\name{read62_01_run_control_params}
\alias{read62_01_run_control_params}
\title{Set the READ62 run control parameters}
\usage{
read62_01_run_control_params(read62_inp = "read62_template.txt",
  read_data_from_surf_dat = TRUE, ibyr = NULL, ibmo = NULL, ibdy = NULL,
  ibhr = NULL, ibsec = 0, ieyr = NULL, iemo = NULL, iedy = NULL,
  iehr = NULL, iesec = 0, jdat = 2, isub = 2, ifmt = 2, pstop = 500,
  lht = FALSE, ltemp = FALSE, lwd = FALSE, lws = FALSE, lxtop = TRUE,
  pvtop = 850, lxsfc = TRUE, zvsfc = 200)
}
\arguments{
\item{read62_inp}{the absolute path and filename for the working CALMET input file.}

\item{read_data_from_surf_dat}{an option to read the time variable data from an extant SURF.DAT file in the working folder.}

\item{ibyr}{the starting year for the CALMET run.}

\item{ibmo}{the starting month for the CALMET run.}

\item{ibdy}{the starting day for the CALMET run.}

\item{ibhr}{the starting hour for the CALMET run.}

\item{ibsec}{the starting second for the CALMET run.}

\item{ieyr}{the ending year for the CALMET run.}

\item{iemo}{the ending month for the CALMET run.}

\item{iedy}{the ending day for the CALMET run.}

\item{iehr}{the ending hour for the CALMET run.}

\item{iesec}{the ending second for the CALMET run.}

\item{jdat}{the type of NCDC input sounding data file; where '1' is the TD-6201 format and '2' is the NCDC FSL format.}

\item{isub}{the format of substitute UP.DAT input sounding data file; where '0' indicates that no substitute will be used, '1' states that the delimiter between sounding levels is a forward slash (and WS and WD have integer representations), and '2' states that the delimiter between sounding levels is a comma (and WS and WD have floating point representations).}

\item{ifmt}{the format of the main UP.DAT input sounding data file; where '1' states that the delimiter between sounding levels is a forward slash (and WS and WD have integer representations), and '2' states that the delimiter between sounding levels is a comma (and WS and WD have floating point representations).}

\item{pstop}{the top pressure level (in mb units) for which data are extracted. The pressure level must correspond to a height that equals or exceeds the top of the CALMET modeling domain, or else CALMET will stop with an error message.}

\item{lht}{a missing data control option for height that is used determine when a sounding level is rejected. If the height is missing from a level, that level will be rejected.}

\item{ltemp}{a missing data control option for temperature that is used determine when a sounding level is rejected. If the temperature is missing from a level, that level will be rejected.}

\item{lwd}{a missing data control option for wind direction that is used determine when a sounding level is rejected. If the wind direction is missing from a level, that level will be rejected.}

\item{lws}{a missing data control option for wind speed that is used determine when a sounding level is rejected. If the wind speed is missing from a level, that level will be rejected.}

\item{lxtop}{choice of whether to extrapolate to extend missing profile data to PSTOP pressure level.}

\item{pvtop}{if 'lxtop' is TRUE, then pvtop is the pressure level corresponding to where valid data must exist.}

\item{lxsfc}{choice of whether to extrapolate to extend missing profile data to the surface.}

\item{zvsfc}{if 'lxsfc' is TRUE, then zvsfc is the height (in meters) corresponding to where valid data must exist.}
}
\description{
This function validates and writes READ62 run control parameters to the working READ62.INP file.
}

