% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read62_inp_generate_template.R
\name{read62_inp_generate_template}
\alias{read62_inp_generate_template}
\title{Generate a template for a READ62 input file}
\usage{
read62_inp_generate_template()
}
\description{
Generate a template for a READ62 input file.
}

