% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replace_in_inp}
\alias{replace_in_inp}
\title{Replace keyword values in CALPUFF system input files}
\usage{
replace_in_inp(calmet_inp_working, keyword, replacement)
}
\arguments{
\item{calmet_inp_working}{the string vector that represents the working CALMET input file.}

\item{keyword}{a string vector of input file keywords for the appropriate section.}

\item{replacement}{a vector of objects that represent formatted strings (for placement in the input file).}
}
\description{
This function allows for the replacement of values in a CALPUFF system input file.
}

