\name{UA.dataprocess}
\alias{UA.dataprocess}
\title{

}

\description{
A work function that transforms df from 'UA.stationDL' to a list.
}

\usage{
UA.dataprocess()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
