\name{UA.stationdata}
\alias{UA.stationdata}
\title{

}

\description{
Obtain a local copy of the available UA data on the NOAA DB.
}

\usage{
UA.stationdata()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
