\name{UA.stationDL}
\alias{UA.stationDL}
\title{

}

\description{
Download the selected data and produce a data frame.
}

\usage{
UA.stationDL()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
