\name{UA.stationselect}
\alias{UA.stationselect}
\title{

}

\description{
Filter and select the list of sites and select one for analysis.
}

\usage{
UA.stationselect()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
