\name{UA.updatcreate}
\alias{UA.updatcreate}
\title{

}

\description{
Transforms list from 'UA.dataprocess' into an UP.DAT file.
}

\usage{
UA.updatcreate()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
