\name{UA.updatfinalize}
\alias{UA.updatfinalize}
\title{

}

\description{
Presents UP.DAT issues and possible strategies to finalize.
}

\usage{
UA.updatfinalize()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
