\name{UA.updatmodify}
\alias{UA.updatmodify}
\title{

}

\description{
Apply a finalization strategy to portions of the UP.DAT file.
}

\usage{
UA.updatmodify()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
