\name{UA.updattest}
\alias{UA.updattest}
\title{

}

\description{
Tests UP.DAT file from 'UA.updatcreate' and logs issues.
}

\usage{
UA.updattest()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
