% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vector_values_to_row_major_strings}
\alias{vector_values_to_row_major_strings}
\title{Transform vector values into strings formatted for GEO.DAT files}
\usage{
vector_values_to_row_major_strings(values_vector, number_cells_across_x,
  number_cells_across_y)
}
\arguments{
\item{values_vector}{a numeric vector}

\item{number_cells_across_x}{the number of grid cells horizontally across the CALMET domain.}

\item{number_cells_across_y}{the number of grid cells vertically across the CALMET domain.}
}
\description{
Transform vector values into strings formatted for GEO.DAT files
}

