\name{wind_opts_params.set}
\alias{wind_opts_params.set}
\title{

}

\description{
Set wind field options and parameters.
}

\usage{
wind_opts_params.set()
}

\arguments{
  \item{ }{

}
  \item{ }{

}
  \item{ }{

}
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

\examples{

}
