% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_01_temporal_params}
\alias{calmet_01_temporal_params}
\title{Set CALMET temporal and model run parameters}
\usage{
calmet_01_temporal_params(calmet_inp = "calmet_template.txt",
  read_data_from_surf_dat = TRUE, ibyr = NULL, ibmo = NULL, ibdy = NULL,
  ibhr = NULL, ibsec = 0, ieyr = NULL, iemo = NULL, iedy = NULL,
  iehr = NULL, iesec = 0, abtz = NULL, nsecdt = 3600, irtype = 1,
  lcalgrd = TRUE, itest = 2, mreg = 0)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}

\item{read_data_from_surf_dat}{an option to read the time variable data from an extant SURF.DAT file in the working folder.}

\item{ibyr}{the starting year for the CALMET run.}

\item{ibmo}{the starting month for the CALMET run.}

\item{ibdy}{the starting day for the CALMET run.}

\item{ibhr}{the starting hour for the CALMET run.}

\item{ibsec}{the starting second for the CALMET run.}

\item{ieyr}{the ending year for the CALMET run.}

\item{iemo}{the ending month for the CALMET run.}

\item{iedy}{the ending day for the CALMET run.}

\item{iehr}{the ending hour for the CALMET run.}

\item{iesec}{the ending second for the CALMET run.}

\item{abtz}{the time zone for the CALMET domain.}

\item{nsecdt}{the number of seconds between time steps.}

\item{irtype}{the CALMET run type, where the '0' option computes wind fields only and the '1' option computes wind fields and several micrometeorological variables (U*, W*, L, Zi, etc.).}

\item{lcalgrd}{the option of whether to compute special data fields required by CALGRID.}

\item{itest}{a flag to stop the model run after the setup phase, which is useful for checking for correctness in model inputs and associated files.}

\item{mreg}{choice of whether to use EPA regulatory options.}
}
\description{
This function validates and writes CALMET parameters for the model starting and ending times, the time zone, and model timings.
}

