% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_5_wind_field_opts_params}
\alias{calmet_5_wind_field_opts_params}
\title{Set CALMET parameters for wind field options and other options}
\usage{
calmet_5_wind_field_opts_params(calmet_inp, iwfcod, ifradj, ikine, iobr, islope,
  iextrp, icalm, bias, rmin2, iprog, isteppg, igfmet, lvary, rmax1, rmax2,
  rmax3, rmin, terrad, r1, r2, rprog, divlim, niter, nsmth, nintr2, critfn,
  alpha, fextr2, nbar, kbar, xbbar, ybbar, xebar, yebar, idiopt1, isurft,
  idiopt2, iupt, zupt, idiopt3, iupwnd, zupwnd, idiopt4, idiopt5, llbreze, nbox,
  xg1, xg2, yg1, yg2, xbcst, ybcst, xecst, yecst, nlb, metbxid)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}

\item{iwfcod}{for model, choose betwen: (0) objective analysis only, or (1) diagnostic wind module.}

\item{ifradj}{compute Froude number adjustment effects.}

\item{ikine}{compute kinematic effects.}

\item{iobr}{use O'Brien procedure for adjustment of the vertical velocity.}

\item{islope}{compute slope flow effects.}

\item{iextrp}{extrapolate surface wind observations to upper layers.}

\item{icalm}{extrapolate surface winds even if calm?}

\item{bias}{layer-dependent biases modifying the weights of surface and upper air stations}

\item{rmin2}{minimum distance from nearest upper air station to surface station for which extrapolation of surface winds at surface station will be allowed.}

\item{iprog}{use gridded prognostic wind field model output fields as input to the diagnostic wind field model.}

\item{isteppg}{timestep in hours of the prognostic model input data.}

\item{igfmet}{use coarse CALMET fields as initial guess fields.}

\item{lvary}{use varying radius of influence.}

\item{rmax1}{maximum radius of influence over land in the surface layer.}

\item{rmax2}{maximum radius of influence over land aloft.}

\item{rmax3}{maximum radius of influence over water.}

\item{rmin}{minimum radius of influence used in the wind field interpolation.}

\item{terrad}{radius of influence of terrain features.}

\item{r1}{the relative weighting of the first guess field and observations in the surface layer.}

\item{r2}{the relative weighting of the first guess field and observations in layers aloft.}

\item{rprog}{relative weighting parameter of the prognostic wind field data.}

\item{divlim}{maximum acceptable divergence in the divergence minimization procedure.}

\item{niter}{maximum number of iterations in the divergence min. procedure.}

\item{nsmth}{number of passes in the smoothing procedure.}

\item{nintr2}{maximum number of stations used in each layer for the interpolation of data to a grid point.}

\item{critfn}{}

\item{alpha}{}

\item{fextr2}{}

\item{nbar}{}

\item{kbar}{}

\item{xbbar}{}

\item{ybbar}{}

\item{xebar}{}

\item{yebar}{}

\item{idiopt1}{}

\item{isurft}{}

\item{idiopt2}{}

\item{iupt}{}

\item{zupt}{}

\item{idiopt3}{}

\item{iupwnd}{}

\item{zupwnd}{}

\item{idiopt4}{}

\item{idiopt5}{}

\item{llbreze}{}

\item{nbox}{}

\item{xg1}{}

\item{xg2}{}

\item{yg1}{}

\item{yg2}{}

\item{xbcst}{}

\item{ybcst}{}

\item{xecst}{}

\item{yecst}{}

\item{nlb}{}

\item{metbxid}{}
}
\description{
This function validates and writes CALMET parameters for wind field options and parameters to the working CALMET.INP file.
}

