% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calmet_define_geophys.R
\name{calmet_define_geophys}
\alias{calmet_define_geophys}
\title{Define the CALMET domain and generate a geophysical input file}
\usage{
calmet_define_geophys(location_name, lat_dec_deg = NULL, lon_dec_deg = NULL,
  lat_lon_grid_loc = 1, domain_width_m = NULL, domain_height_m = NULL,
  cell_resolution_m = 250, four_seasons = TRUE,
  four_season_breaks = c("03-15", "05-31", "08-31", "11-15"),
  LU_method = "MODIS_Global", shapefile_dir = NULL)
}
\arguments{
\item{location_name}{an assigned name for the CALMET domain.}

\item{lat_dec_deg}{the latitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_lon_grid_loc argument.}

\item{lon_dec_deg}{the longitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_lon_grid_loc argument.}

\item{lat_lon_grid_loc}{the location of the lat/long inputs in relation to the domain. Choices are: 1 (center), 2 (lower left), 3 (lower right), 4 (upper left), 5 (upper right).}

\item{domain_width_m}{the desired width of the meteorological domain in meters.}

\item{domain_height_m}{the desired height of the meteorological domain in meters.}

\item{cell_resolution_m}{the desired cell resolution in meters.}

\item{four_seasons}{an option for whether to split the generated GEO.DAT into four separate files, one for each of four seasons.}

\item{four_season_breaks}{a character vector (in the format MM-DD) that specifies when the different seasons are to transition to the next.}

\item{LU_method}{the method for obtaining CALMET land use codes for the selected domain; current options are (1) "MODIS_Global" and (2) "BTM_BC" (a method specific to the province of British Columbia in Canada, and, requires a local copy of the BC BTM V1 shapefile).}

\item{shapefile_dir}{if the chosen 'LU_method' requires the use of a local shapefile (currently, the "BTM_BC" uses shapefile data), supply the relative or absolute path to the shapefile here.}
}
\description{
Define the CALMET domain and determine the best gridded values for land use, terrain heights, and micrometeorological parameters for creation of a geophysical input file.
}
\examples{
\dontrun{
# Generate a geophysical input file (GEO.DAT) for CALMET
calmet_define_geophys(location_name = "test",
                      lat_dec_deg = 49.196116,
                      lon_dec_deg = -122.505866,
                      lat_lon_grid_loc = 1,
                      domain_width_m = 117000,
                      domain_height_m = 43250,
                      cell_resolution_m = 500)
}
}

