% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_define_geophys}
\alias{calmet_define_geophys}
\title{Define the CALMET domain and generate a geophysical input file}
\usage{
calmet_define_geophys(location_name, lat_dec_deg = NULL, lon_dec_deg = NULL,
  lat_lon_grid_loc = 1, domain_width_m = NULL, domain_height_m = NULL,
  cell_resolution_m = 250, four_seasons = TRUE,
  four_season_breaks = c("03-15", "05-31", "08-31", "11-15"),
  download_SRTM = TRUE, SRTM_file_path = NULL, LU_method = "MODIS_Global",
  shapefile_dir = NULL)
}
\arguments{
\item{location_name}{an assigned name for the CALMET domain.}

\item{lat_dec_deg}{the latitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_lon_grid_loc argument.}

\item{lon_dec_deg}{the longitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_lon_grid_loc argument.}

\item{lat_lon_grid_loc}{the location of the lat/long inputs in relation to the domain. Choices are: 1 (center), 2 (lower left), 3 (lower right), 4 (upper left), 5 (upper right).}

\item{domain_width_m}{the desired width of the meteorological domain in meters.}

\item{domain_height_m}{the desired height of the meteorological domain in meters.}

\item{cell_resolution_m}{the desired cell resolution in meters.}

\item{download_SRTM}{a choice of whether to download the SRTM GeoTIFF height data from a server or read the identical files from a local folder.}

\item{SRTM_file_path}{path to a folder containing a collection of SRTM V4 zip archive files.}

\item{LU_method}{the method for obtaining CALMET land use codes for the selected domain; current options are (1) "MODIS_Global" and (2) "BTM_BC" (a method specific to the province of British Columbia in Canada, and, requires a local copy of the BC BTM V1 shapefile).}

\item{shapefile_dir}{if the chosen 'LU_method' requires the use of a local shapefile, supply the relative or absolute path to the shapefile here.}
}
\description{
Define the CALMET domain and determine the best gridded values for land use, terrain heights, and micrometeorological parameters for creation of a geophysical input file.
}
\examples{
\dontrun{
# Create a CALMET domain of 100 by 100 km in the Los Angeles area.
# Chosen lat/lon coordinates are for the center of the domain.
calmet_define_geophys(location_name = "losangeles",
                      lat_dec_deg = 34.050184,
                      lon_dec_deg = -118.253959,
                      lat_lon_grid_loc = 1,
                      domain_width_m = 100000,
                      domain_height_m = 100000,
                      cell_resolution_m = 250
                      download_SRTM = TRUE)
}
}

