% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calmet_get_ncdc_station_data.R
\name{calmet_get_ncdc_station_data}
\alias{calmet_get_ncdc_station_data}
\title{Obtain NCDC station data}
\usage{
calmet_get_ncdc_station_data(data_filename = NULL, local_archive_dir = NULL,
  year = NULL, bbox_lat_lon = NULL)
}
\arguments{
\item{data_filename}{a string representing the exact filename for the archive to be retrieved.}

\item{local_archive_dir}{a local path containing an archive of gzipped NCDC station data files.}

\item{year}{the year of the data archive.}

\item{bbox_lat_lon}{a spatial bounding box in projected in lat/lon coordinates.}
}
\description{
Obtain NCDC station data for all stations residing in the CALMET domain during a specied time period.
}
\examples{
\dontrun{
# Get 2010 station data for a previously defined domain
calmet_get_ncdc_station_data(year = 2010,
                             bbox_lat_lon = bbox)
}
}

