% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_inp_finalize}
\alias{calmet_inp_finalize}
\title{Finalize and generate CALMET input files}
\usage{
calmet_inp_finalize(calmet_inp = "calmet_template.txt",
  get_filenames_from_wd = TRUE, geodat = NULL, srfdat = NULL,
  clddat = NULL, prcdat = NULL, wtdat = NULL, metlst = NULL,
  metdat = NULL, pacdat = NULL, lcfiles = TRUE, nusta = NULL,
  nowsta = NULL, nm3d = NULL, nigf = NULL, updat = NULL,
  seadat = NULL, m3ddat = NULL, igfdat = NULL, diadat = NULL,
  prgdat = NULL, tstprt = NULL, tstout = NULL, tstkin = NULL,
  tstfrd = NULL, tstslp = NULL, dcstgd = NULL)
}
\arguments{
\item{calmet_inp}{the absolute path and filename for the working CALMET input file.}

\item{get_filenames_from_wd}{}

\item{geodat}{}

\item{srfdat}{}

\item{clddat}{}

\item{prcdat}{}

\item{wtdat}{}

\item{metlst}{}

\item{metdat}{}

\item{pacdat}{}

\item{lcfiles}{}

\item{nusta}{}

\item{nowsta}{}

\item{nm3d}{}

\item{nigf}{}

\item{updat}{}

\item{seadat}{}

\item{m3ddat}{}

\item{igfdat}{}

\item{diadat}{}

\item{prgdat}{}

\item{tstprt}{}

\item{tstout}{}

\item{tstkin}{}

\item{tstfrd}{}

\item{tstslp}{}

\item{dcstgd}{}
}
\description{
This function sets the input and output filenames for CALMET model runs generates one or a series of CALMET.INP files.
}

