% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_landuse_BTM_BC}
\alias{calmet_landuse_BTM_BC}
\title{Generate a vector of CALMET landuse categories from BC shapefile data}
\usage{
calmet_landuse_BTM_BC(location_name, bbox_longlat, bbox_UTM, UTM_zone,
  UTM_hemisphere)
}
\arguments{
\item{location_name}{an assigned name for the CALMET domain.}

\item{bbox_longlat}{}

\item{bbox_UTM}{}

\item{UTM_zone}{}

\item{UTM_hemisphere}{}
}
\description{
Generate a vector of CALMET landuse categories from BC shapefile data.
}

