% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calmet_plot_domain.R
\name{calmet_plot_domain}
\alias{calmet_plot_domain}
\title{Plot met stations in the CALMET domain}
\usage{
calmet_plot_domain()
}
\description{
Plot all surface meteorological stations in CALMET domain
}
\examples{
\dontrun{
# Plot all stations in NCDC database in the defined CALMET domain
calmet_plot_domain()
}
}

