% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calmet_surface_met}
\alias{calmet_surface_met}
\title{Obtain surface meteorology and generate input files}
\usage{
calmet_surface_met(start_year, end_year, lat_dec_deg = NULL,
  lon_dec_deg = NULL, lat_lon_grid_loc = 1, domain_width_m = NULL,
  domain_height_m = NULL, time_offset, output_file = "surf.dat")
}
\arguments{
\item{start_year}{the starting year for the SURF.DAT file.}

\item{end_year}{the ending year for the SURF.DAT file.}

\item{lat_dec_deg}{the latitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_long_grid_loc argument.}

\item{lon_dec_deg}{the longitude of the CALMET domain in decimal degrees. The location of this point is defined in the lat_long_grid_loc argument.}

\item{lat_lon_grid_loc}{the location of the lat/long inputs in relation to the domain. Choices are: 1 (center), 2 (lower left), 3 (lower right), 4 (upper left), 5 (upper right).}

\item{domain_width_m}{the width of the meteorological domain in meters.}

\item{domain_height_m}{the height of the meteorological domain in meters.}

\item{output_file}{the specified filename for the SURF.DAT file. Defaults to "surf.dat".}
}
\description{
Obtain surface meteorology and generate input files
}
\examples{
\dontrun{
# Generate "surf.dat" from previously defined period
calmet_surface_met()
}
}

