% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_01_run_control_params}
\alias{calpuff_01_run_control_params}
\title{Set CALPUFF run control parameters}
\usage{
calpuff_01_run_control_params(calpuff_inp = "calpuff_template.txt",
  read_data_from_surf_dat = TRUE, ibyr = NULL, ibmo = NULL, ibdy = NULL,
  ibhr = NULL, ibsec = 0, ieyr = NULL, iemo = NULL, iedy = NULL,
  iehr = NULL, iesec = 0, abtz = NULL, nsecdt = 3600, nse, itest,
  mrestart, nrespd, metfm, mprffm, avet, pgtime, ioutu, iovers)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{metrun}{}

\item{ibyr}{the starting year for the CALPUFF run.}

\item{ibmo}{the starting month for the CALPUFF run.}

\item{ibdy}{the starting day for the CALPUFF run.}

\item{ibhr}{the starting hour for the CALPUFF run.}

\item{ibsec}{the starting second for the CALPUFF run.}

\item{ieyr}{the ending year for the CALPUFF run.}

\item{iemo}{the ending month for the CALPUFF run.}

\item{iedy}{the ending day for the CALPUFF run.}

\item{iehr}{the ending hour for the CALPUFF run.}

\item{iesec}{the ending second for the CALPUFF run.}

\item{abtz}{the time zone for the CALPUFF domain.}

\item{nsecdt}{the number of seconds between time steps.}

\item{nse}{number of stationary emission sources.}

\item{itest}{a flag to stop the model run after the setup phase, which is useful for checking for correctness in model inputs and associated files.}

\item{mrestart}{}

\item{nrespd}{}

\item{metfm}{}

\item{mprffm}{}

\item{avet}{}

\item{pgtime}{}

\item{ioutu}{}

\item{iovers}{}
}
\description{
This function validates and writes CALPUFF run control parameters.
}

