% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_04_grid_levels}
\alias{calpuff_04_grid_levels}
\title{Set CALPUFF parameters for the map projection, datum, and grid definitions}
\usage{
calpuff_04_grid_levels(calpuff_inp, pmap = "UTM", feast = 0, fnorth = 0,
  iutmzn = NULL, utmhem = NULL, rlat0 = "40N", rlon0 = "90W",
  xlat1 = "30N", xlat2 = "60N", datum = "WGS-84", nx = NULL,
  ny = NULL, dgridkm = NULL, xorigkm = NULL, yorigkm = NULL, nz = 12,
  zface = c(0, 20, 40, 80, 100, 150, 200, 300, 400, 800, 1400, 2000, 3000),
  ibcomp, jbcomp, iecomp, jecomp, lsamp, ibsamp, jbsamp, iesamp, jesamp, meshdn)
}
\arguments{
\item{calpuff_inp}{}

\item{pmap}{}

\item{feast}{}

\item{fnorth}{}

\item{iutmzn}{}

\item{utmhem}{}

\item{rlat0}{}

\item{rlon0}{}

\item{xlat1}{}

\item{xlat2}{}

\item{datum}{}

\item{nx}{}

\item{ny}{}

\item{dgridkm}{}

\item{xorigkm}{}

\item{yorigkm}{}

\item{nz}{}

\item{zface}{}

\item{ibcomp}{}

\item{jbcomp}{}

\item{iecomp}{}

\item{jecomp}{}

\item{lsamp}{}

\item{ibsamp}{}

\item{jbsamp}{}

\item{iesamp}{}

\item{jesamp}{}

\item{meshdn}{}
}
\description{
This function validates and writes CALPUFF parameters for the map projection, datum, and grid definitions to the working CALPUFF.INP file.
}

