% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_05_output_opts}
\alias{calpuff_05_output_opts}
\title{Set the CALPUFF output options}
\usage{
calpuff_05_output_opts(calpuff_inp = "calpuff_template.txt", icon, idry, iwet,
  it2d, irho, ivis, lcomprs, iqaplot, ipftrak, imflx, imbal, inrise, icprt,
  idprt, iwprt, icfrq, idfrq, iwfrq, iprtu, imesg, ldebug, ipfdeb, npfdeb, nn1,
  nn2)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{icon}{}

\item{idry}{}

\item{iwet}{}

\item{it2d}{}

\item{irho}{}

\item{ivis}{}

\item{lcomprs}{}

\item{iqaplot}{}

\item{ipftrak}{}

\item{imflx}{}

\item{imbal}{}

\item{inrise}{}

\item{icprt}{}

\item{idprt}{}

\item{iwprt}{}

\item{icfrq}{}

\item{idfrq}{}

\item{iwfrq}{}

\item{iprtu}{}

\item{imesg}{}

\item{ldebug}{}

\item{ipfdeb}{}

\item{npfdeb}{}

\item{nn1}{}

\item{nn2}{}
}
\description{
This function validates and writes CALPUFF output options.
}

