% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_06_complex_terrain_inputs}
\alias{calpuff_06_complex_terrain_inputs}
\title{Set the CALPUFF subgrid scale complex terrain inputs}
\usage{
calpuff_06_complex_terrain_inputs(calpuff_inp = "calpuff_template.txt",
  nhill = 0, nctrec = 0, mhill = 2, xhill2m = 1, zhill2m = 1,
  xctdmkm = 0, yctdmkm = 0)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{nhill}{the number of terrain features.}

\item{nctrec}{the number of special complex terrain receptors.}

\item{mhill}{}

\item{xhill2m}{}

\item{zhill2m}{}

\item{xctdmkm}{}

\item{yctdmkm}{}
}
\description{
This function validates and writes CALPUFF subgrid scale complex terrain inputs.
}

