% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calpuff_06_complex_terrain_inputs.R
\name{calpuff_06_complex_terrain_inputs}
\alias{calpuff_06_complex_terrain_inputs}
\title{Set the CALPUFF subgrid scale complex terrain inputs}
\usage{
calpuff_06_complex_terrain_inputs(calpuff_inp = "calpuff_template.txt",
  nhill = 0, nctrec = 0, mhill = 2, xhill2m = 1, zhill2m = 1,
  xctdmkm = 0, yctdmkm = 0)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{nhill}{the number of terrain features.}

\item{nctrec}{the number of special complex terrain receptors.}

\item{mhill}{provenance of terrain and CTSG Receptor data: (1) hill and receptor data created by CTDM processors and read from HILL.DAT and HILLRCT.DAT files, or (2) hill data created by OPTHILL.}

\item{xhill2m}{factor to convert horizontal dimensions to meters.}

\item{zhill2m}{factor to convert vertical dimensions to meters.}

\item{xctdmkm}{the x-origin of the CTDM system relative to the CALPUFF coordinate system, in kilometers.}

\item{yctdmkm}{the y-origin of the CTDM system relative to the CALPUFF coordinate system, in kilometers.}
}
\description{
This function validates and writes CALPUFF subgrid scale complex terrain inputs.
}

