% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_08_size_ddep_particles}
\alias{calpuff_08_size_ddep_particles}
\title{Set the CALPUFF size parameters for dry deposition of particles}
\usage{
calpuff_08_size_ddep_particles(calpuff_inp,
  species_geom_mass_mean_diam = NULL, species_geom_SD = NULL)
}
\arguments{
\item{calpuff_inp}{}

\item{species_geom_mass_mean_diam}{}

\item{species_geom_SD}{}
}
\description{
This function validates and writes CALPUFF size parameters for dry deposition of particles.
}

