% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_09_misc_ddep}
\alias{calpuff_09_misc_ddep}
\title{Set miscellaneous CALPUFF parameters for dry deposition}
\usage{
calpuff_09_misc_ddep(calpuff_inp = "calpuff_template.txt", rcutr, rgr, reactr,
  nint, iveg)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{rcutr}{}

\item{rgr}{}

\item{reactr}{}

\item{nint}{}

\item{iveg}{}
}
\description{
This function validates and writes miscellaneous CALPUFF parameters for dry deposition.
}

