% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_10_wdep_params}
\alias{calpuff_10_wdep_params}
\title{Set the CALPUFF parameters for wet deposition}
\usage{
calpuff_10_wdep_params(calpuff_inp = "calpuff_template.txt",
  species_names = NULL, species_scaveng_coeff_liquid_precip = NULL,
  species_scaveng_coeff_froze_precip = NULL)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{species_names}{a vector list of species for which wet deposition parameters will be supplied.}

\item{species_scaveng_coeff_liquid_precip}{}

\item{species_scaveng_coeff_froze_precip}{}
}
\description{
This function validates and writes CALPUFF parameters for wet deposition.
}

