% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_11_chem_params}
\alias{calpuff_11_chem_params}
\title{Set the CALPUFF chemistry parameters}
\usage{
calpuff_11_chem_params(calpuff_inp = "calpuff_template.txt", moz, bcko3, mnh3,
  mavgnh3, bcknh3, rnite1, rnite2, rnite3, mh2o2, bckh2o2, bckpmf, ofrac, vcnx,
  ndecay, spec1, spec2)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{moz}{}

\item{bcko3}{}

\item{mnh3}{}

\item{mavgnh3}{}

\item{bcknh3}{}

\item{rnite1}{}

\item{rnite2}{}

\item{rnite3}{}

\item{mh2o2}{}

\item{bckh2o2}{}

\item{bckpmf}{}

\item{ofrac}{}

\item{vcnx}{}

\item{ndecay}{}

\item{spec1}{}

\item{spec2}{}
}
\description{
This function validates and writes CALPUFF parameters for atmospheric chemistry.
}

