% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_12_misc_disp_comp_params}
\alias{calpuff_12_misc_disp_comp_params}
\title{Set miscellaneous CALPUFF dispersion and computational parameters}
\usage{
calpuff_12_misc_disp_comp_params(calpuff_inp = "calpuff_template.txt",
  sytdep = 550, mhftsz = FALSE, jsup = 5, conk1 = 0.01, conk2 = 0.1,
  tbd = 0.5, iurb1 = 10, iurb2 = 19, ilanduin = 20, z0in = 0.25,
  xlaiin = 0, elevin = 0, xlatin = 0, xlonin = 0, anemht = 10,
  isigmav = 1, imixctdm = 0, xmxlen = 1, xsamlen = 1, mxnew = 99,
  mxsam = 99, ncount = 2, symin = 1, szmin = 1, szcap_m = 5e+06,
  svmin = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.37, 0.37, 0.37, 0.37, 0.37, 0.37),
  swmin = c(0.2, 0.12, 0.08, 0.06, 0.03, 0.016, 0.2, 0.12, 0.08, 0.06, 0.03,
  0.016), cdiv = c(0, 0), nlutibl = 4, wscalm = 0.5, xmaxzi = 3000,
  xminzi = 20, wscat = c(1.54, 3.09, 5.14, 8.23, 10.8), plx0 = c(0.07,
  0.07, 0.1, 0.15, 0.35, 0.55), ptg0 = c(0.02, 0.035), ppc = c(0.5, 0.5,
  0.5, 0.5, 0.35, 0.35), sl2pf = 10, nsplit = 3, iresplit = c(0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0), zisplit = 100,
  roldmax = 0.25, nsplith = 5, sysplith = 1, shsplith = 2,
  cnsplith = 1e-07, epsslug = 1e-04, epsarea = 1e-06, dsrise = 1,
  htminbc = 500, rsampbc = 10, mdepbc = TRUE)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{sytdep}{the horizontal size of puff (in units of meters) beyond which the time-dependent Heffter dispersion equations are used to determine sigma-y and sigma-z.}

\item{mhftsz}{the choice of whether to use the Heffter equation to calculate sigma z values.}

\item{jsup}{the stability class used to determine plume growth rates for puffs above the boundary layer.}

\item{conk1}{the vertical dispersion constant for stable conditions.}

\item{conk2}{the vertical dispersion constant for neutral and unstable conditions.}

\item{tbd}{the factor for determining the transition-point from Schulman-Scire to Huber-Snyder Building downwash schemes.}

\item{iurb1}{the lower range of land use categories for which urban dispersion is assumed.}

\item{iurb2}{the upper range of land use categories for which urban dispersion is assumed.}

\item{ilanduin}{the land use category for the modeling domain. Used for site characterization parameters with single-point Met data files and ignored unless metfm is either 2, 3, 4, or 5.}

\item{z0in}{the roughness length (in units of meters) for the modeling domain. Used for site characterization parameters with single-point Met data files and ignored unless metfm is either 2, 3, 4, or 5.}

\item{xlaiin}{the leaf area index for the modeling domain. Used for site characterization parameters with single-point Met data files and ignored unless metfm is either 2, 3, 4, or 5.}

\item{elevin}{the elevation of the site (in units of meters ASL). Used for site characterization parameters with single-point Met data files and ignored unless metfm is either 2, 3, 4, or 5.}

\item{xlatin}{the latitude of the site (in decimal degrees). Used for site characterization parameters with single-point Met data files and ignored unless metfm is either 2, 3, 4, or 5.}

\item{xlonin}{the longitude of the site (in decimal degrees). Used for site characterization parameters with single-point Met data files and ignored unless metfm is either 2, 3, 4, or 5.}

\item{anemht}{the anemometer height at the site (in units of meters ASL). Used for site characterization parameters with single-point Met data files and ignored unless metfm is either 2 or 3.}

\item{isigmav}{the choice of which lateral turbulance data in PROFILE.DAT file to use: (0) read sigma-theta, or (1) read sigma-v. Used for site characterization parameters with single-point Met data files and ignored unless metfm is either 4 or 5, and mturbvw is either 1 or 3.}

\item{imixctdm}{the choice of method for reading mixing heights: (0) read predicted, or (1) read observed. Used for site characterization parameters with single-point Met data files and ignored unless metfm is 4.}

\item{xmxlen}{the maximum length of a slug (in meteorological grid cell units).}

\item{xsamlen}{the maximum travel distance of a slug or puff (in meteorological grid cell units) during each time step.}

\item{mxnew}{the maximum number of slugs or puffs released from one source during each time step.}

\item{mxsam}{the maximum number of sampling steps for one puff or slug during each time step.}

\item{ncount}{the number of iterations used when computing the transport wind for a sampling step that includes gradual rise.}

\item{symin}{the minimum sigma y (in units of meters) for a new puff or slug.}

\item{szmin}{the minimum sigma z (in units of meters) for a new puff or slug.}

\item{szcap_m}{the maximum sigma z (in units of meters) that is permitted.}

\item{svmin}{the default minimum turbulence velocities sigma v (in units of meters per second) for each stability class over land and over water.}

\item{swmin}{the default minimum turbulence velocities sigma w (in units of meters per second) for each stability class over land and over water.}

\item{cdiv}{the divergence criterion for dw/dz across the puff; used to initiate adjustment for horizontal convergence.}

\item{nlutibl}{the search radius (in meteorological grid cell units) for nearest land and water cells.}

\item{wscalm}{the minimum wind speed (in units of meters per second) allowed for non-calm conditions.}

\item{xmaxzi}{the maximum mixing height (in units of meters).}

\item{xminzi}{the minimum mixing height (in units of meters).}

\item{wscat}{the upper-bound wind-speeds (in units of meters per second) for the the default wind speed classes.}

\item{plx0}{the default wind speed profile power-law exponent values for stability classes 1 through 6.}

\item{ptg0}{the default potential temperature gradient values for stable classes E and F.}

\item{ppc}{the default plume path coefficients for each stability class (ignored unless the option for partial plume height terrain adjustment, mctadj, is selected by using the value 3).}

\item{sl2pf}{the slug-to-puff transition criterion factor equal to sigma-y or length of a slug.}

\item{nsplit}{the number of puffs that result every time a puff is split vertically.}

\item{iresplit}{the time(s) of a day when split puffs are eligible to be split vertically once again.}

\item{zisplit}{the minimum mixing height (in units of meters) for which vertical puff-splitting is allowed in the subsequent hour.}

\item{roldmax}{the maximum value of ratio of the mixing height to the maximum mixing height experienced by the puff for which vertical puff-splitting is allowed in the subsequent hour.}

\item{nsplith}{the number of puffs that result every time a puff is split horizontally.}

\item{sysplith}{the minimum sigma-y (in meteorological grid cell units) of puff before horizontal puff-splitting is allowed.}

\item{shsplith}{the minimum puff elongation rate (sysplith units per hour) due to wind shear before horizontal puff-splitting is allowed.}

\item{cnsplith}{the minimum concentration (in units of grams per cubic meter) of each species within a puff before it may be split horizontally. Provide vector of length corresponding to the number of species; if a single value is provided, it will be applied to all species.}

\item{epsslug}{the fractional convergence criterion for numerical slug sampling integration.}

\item{epsarea}{the fractional convergence criterion for numerical area source integration.}

\item{dsrise}{the trajectory step-length (in units of meters) used for numerical rise integration.}

\item{htminbc}{the minimum height (in units of meters) to which boundary condition puffs are mixed as they are emitted (ignored unless mbcon is 2).}

\item{rsampbc}{the search radious (in units of km) about a receptor for sampling the nearest boundary condition puff.}

\item{mdepbc}{the choice of whether to perform a near-surface depletion adjustment to the concentration profile when sampling boundary condition puffs.}
}
\description{
This function validates and writes miscellaneous CALPUFF dispersion and computational parameters.
}

