% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_12_misc_disp_comp_params}
\alias{calpuff_12_misc_disp_comp_params}
\title{Set miscellaneous CALPUFF dispersion and computational parameters}
\usage{
calpuff_12_misc_disp_comp_params(calpuff_inp, sytdep, mhftsz, jsup, conk1,
  conk2, tbd, iurb1, iurb2, ilanduin, z0in, xlaiin, elevin, xlatin, xlonin,
  anemht, isigmav, imixctdm, xmxlen, xsamlen, mxnew, mxsam, ncount, symin,
  szmin, szcap_m, svmin, swmin, cdiv, nlutibl, wscalm, xmaxzi, xminzi, wscat,
  plx0, ptg0, ppc, sl2pf, nsplit, iresplit, zisplit, roldmax, nsplith, sysplith,
  shsplith, cnsplith, epsslug, epsarea, dsrise, htminbc, rsampbc, mdepbc)
}
\arguments{
\item{calpuff_inp}{}

\item{sytdep}{}

\item{mhftsz}{}

\item{jsup}{}

\item{conk1}{}

\item{conk2}{}

\item{tbd}{}

\item{iurb1}{}

\item{iurb2}{}

\item{ilanduin}{}

\item{z0in}{}

\item{xlaiin}{}

\item{elevin}{}

\item{xlatin}{}

\item{xlonin}{}

\item{anemht}{}

\item{isigmav}{}

\item{imixctdm}{}

\item{xmxlen}{}

\item{xsamlen}{}

\item{mxnew}{}

\item{mxsam}{}

\item{ncount}{}

\item{symin}{}

\item{szmin}{}

\item{szcap_m}{}

\item{svmin}{}

\item{swmin}{}

\item{cdiv}{}

\item{nlutibl}{}

\item{wscalm}{}

\item{xmaxzi}{}

\item{xminzi}{}

\item{wscat}{}

\item{plx0}{}

\item{ptg0}{}

\item{ppc}{}

\item{sl2pf}{}

\item{nsplit}{}

\item{iresplit}{}

\item{zisplit}{}

\item{roldmax}{}

\item{nsplith}{}

\item{sysplith}{}

\item{shsplith}{}

\item{cnsplith}{}

\item{epsslug}{}

\item{epsarea}{}

\item{dsrise}{}

\item{htminbc}{}

\item{rsampbc}{}

\item{mdepbc}{}
}
\description{
This function validates and writes miscellaneous CALPUFF dispersion and computational parameters.
}

