% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_13_pt_source_params}
\alias{calpuff_13_pt_source_params}
\title{Set one or more CALPUFF point source parameters}
\usage{
calpuff_13_pt_source_params(calpuff_inp, npt1 = NULL, iptu = NULL,
  nspt1 = NULL, npt2 = NULL, ptsource_name = NULL,
  ptsource_params = NULL, ptsource_zpltfm = NULL, ptsource_fmfac = NULL,
  downwash_output = NULL)
}
\arguments{
\item{calpuff_inp}{}

\item{npt1}{}

\item{iptu}{}

\item{nspt1}{}

\item{npt2}{}

\item{ptsource_name}{}

\item{ptsource_params}{}

\item{ptsource_zpltfm}{}

\item{ptsource_fmfac}{}

\item{downwash_output}{}
}
\description{
This function validates and writes CALPUFF point source parameters.
}

