% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_14_area_source_params}
\alias{calpuff_14_area_source_params}
\title{Set one or more CALPUFF area source parameters}
\usage{
calpuff_14_area_source_params(calpuff_inp, nar1 = NULL, iaru = NULL,
  nsar1 = NULL, nar2 = NULL, arsource_params = NULL,
  arsource_coords = NULL)
}
\arguments{
\item{calpuff_inp}{}

\item{nar1}{}

\item{iaru}{}

\item{nsar1}{}

\item{nar2}{}

\item{arsource_params}{}

\item{arsource_coords}{}
}
\description{
This function validates and writes CALPUFF area source parameters.
}

