% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calpuff_15_line_source_params.R
\name{calpuff_15_line_source_params}
\alias{calpuff_15_line_source_params}
\title{Set one or more CALPUFF line source parameters}
\usage{
calpuff_15_line_source_params(calpuff_inp = "calpuff_template.txt",
  nln2 = NULL, nlines = NULL, ilnu = NULL, nsln1 = NULL,
  mxnseg = NULL, nlrise = NULL, xl = NULL, hbl = NULL, wbl = NULL,
  wml = NULL, dxl = NULL, fprimel = NULL, lnsource_params = NULL)
}
\arguments{
\item{calpuff_inp}{the absolute path and filename for the working CALPUFF input file.}

\item{nln2}{the number of buoyant line sources with variable location and emission parameters.}

\item{nlines}{the number of buoyant line sources.}

\item{ilnu}{the units used for line source emissions, where the possible selections are: (1) "g/s", (2) "kg/hr", (3) "lb/hr", (4) "tons/yr", (5) "Odour Unit * m^3/s", (6) "Odour Unit * m^3/min", and (7) "metric tons/yr".}

\item{nsln1}{the number of source-species combinations with variable emissions scaling factors.}

\item{mxnseg}{the maximum number of line segments to model each line.}

\item{nlrise}{the number of distances at which transitional rise is computed.}

\item{xl}{the average building length in meters.}

\item{hbl}{the average building height in meters.}

\item{wbl}{the average building width in meters.}

\item{wml}{the average line source width in meters.}

\item{dxl}{the average separation between buildings in meters.}

\item{fprimel}{the average buoyancy parameter, in units of m^4/s^3}

\item{lnsource_params}{}
}
\description{
This function validates and writes CALPUFF line source parameters.
}

