% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calpuff_add_line_sources.R
\name{calpuff_add_line_sources}
\alias{calpuff_add_line_sources}
\title{Add line sources to a list for later use in CALPUFF}
\usage{
calpuff_add_line_sources(src_name, species_name, lat_dec_deg = NULL,
  lon_dec_deg = NULL, x_coord_km = NULL, y_coord_km = NULL,
  UTM_zone = NULL, UTM_hemisphere = NULL, release_hgt, base_elev,
  emission_rate, emission_units)
}
\arguments{
\item{src_name}{the name of the source emitting the species.}

\item{species_name}{the name of the species undergoing emissions.}

\item{lat_dec_deg}{a vector of 2 latitude values for the line source in units of decimal degrees.}

\item{lon_dec_deg}{a vector of 2 longitude values for the line source in units of decimal degrees.}

\item{x_coord_km}{a vector of 2 UTM easting values for the line source in km units.}

\item{y_coord_km}{a vector of 2 UTM northing values for the area source in km units.}

\item{UTM_zone}{the UTM zone for the line source.}

\item{UTM_hemisphere}{the UTM hemisphere for the line source.}

\item{release_hgt}{the release height of the line source in meters above ground level (m AGL).}

\item{base_elev}{the ground elevation at the location of the area source in meters above sea level (m ASL).}

\item{emission_rate}{the rate of constant emissions from the line source; units are defined in the 'emission_units' argument.}

\item{emission_units}{the units applied to the value defined in the 'emission_rate' argument. The possible selections are: (1) "g/s", (2) "kg/hr", (3) "lb/hr", (4) "tons/yr", (5) "Odour Unit * m3/s", (6) "Odour Unit * m3/min", (7) "metric tons/yr", (8) "Bq/s", and (9) "GBq/yr".}
}
\description{
Add line sources to a list for later use in CALPUFF
}

