% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_create_varying_area_sources}
\alias{calpuff_create_varying_area_sources}
\title{Create CALPUFF input file with area sources that vary in location and time}
\usage{
calpuff_create_varying_area_sources(CSV_input = NULL, df_input = NULL,
  pollutant_MW)
}
\arguments{
\item{CSV_input}{a path to a CSV file containing hourly area source data.}

\item{df_input}{a data frame containing hourly area source data.}

\item{pollutant_MW}{a vector containing molecular weights for all pollutants in the supplied dataset.}
}
\description{
Create CALPUFF input file with area sources that vary in location and time.
}

