% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calpuff_create_varying_point_sources}
\alias{calpuff_create_varying_point_sources}
\title{Create CALPUFF input file with point sources that vary in location and time}
\usage{
calpuff_create_varying_point_sources(CSV_input = NULL, df_input = NULL,
  src_name, species_name)
}
\arguments{
\item{CSV_input}{a path to a CSV file containing hourly point source data.}

\item{df_input}{a data frame containing hourly point source data.}

\item{src_name}{the name of the source emitting the species.}

\item{species_name}{the name of the species undergoing emissions.}
}
\description{
Create CALPUFF input file with point sources that vary in location and time.
}

