% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download_SRTMV4_GeoTIFF}
\alias{download_SRTMV4_GeoTIFF}
\title{Download an SRTM V4 GeoTIFF file}
\usage{
download_SRTMV4_GeoTIFF(SP_object)
}
\arguments{
\item{SP_object}{a SpatialPoints object with a longlat projection.}
}
\description{
Download an SRTM V4 GeoTIFF file based on lat/lon coordinates
}

