% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download_SRTMV4_GeoTIFF}
\alias{download_SRTMV4_GeoTIFF}
\title{Download an SRTM V4 GeoTIFF file}
\usage{
download_SRTMV4_GeoTIFF(lon, lat, download, path)
}
\arguments{
\item{lon}{a longitude supplied as decimal degrees.}

\item{lat}{a latitude supplied as decimal degrees.}

\item{download}{a choice of whether to download the file if it already exists in the specified local path.}

\item{path}{a path for which GeoTIFF files may exist locally.}
}
\description{
Download an SRTM V4 GeoTIFF file based on lat/lon coordinates
}

