% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.calmet.domain}
\alias{plot.calmet.domain}
\title{Plot met stations in the CALMET domain}
\usage{
\method{plot}{calmet.domain}()
}
\description{
Plot all surface meteorological stations in CALMET domain
}
\examples{
\dontrun{
# Plot all stations in NCDC database in the defined CALMET domain
plot.calmet.domain()
}
}

