% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read62_01_run_control_params}
\alias{read62_01_run_control_params}
\title{Set READ62 run control parameters}
\usage{
read62_01_run_control_params(read62_inp = "read62_template.txt",
  read_data_from_surf_dat = TRUE, ibyr = NULL, ibmo = NULL, ibdy = NULL,
  ibhr = NULL, ibsec = 0, ieyr = NULL, iemo = NULL, iedy = NULL,
  iehr = NULL, iesec = 0, jdat = 2, isub = 2, ifmt = 2, pstop = 500,
  lxtop = TRUE, pvtop = 850, lxsfc = TRUE, zvsfc = 200)
}
\arguments{
\item{read62_inp}{the absolute path and filename for the working CALMET input file.}
}
\description{
This function validates and writes READ62 run control parameters to the working READ62.INP file.
}

