% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replace_in_inp.R
\name{replace_in_inp}
\alias{replace_in_inp}
\title{Replace keyword values in CALPUFF system input files}
\usage{
replace_in_inp(inp_file_working, keyword, replacement)
}
\arguments{
\item{inp_file_working}{the string vector that represents the working input file.}

\item{keyword}{a string vector of input file keywords for the appropriate section.}

\item{replacement}{a vector of objects that represent formatted strings (for placement in the input file).}
}
\description{
This function allows for the replacement of values in a CALPUFF system input file.
}

