% Generated by roxygen2 (4.0.1): do not edit by hand
\name{surf.dat.generate}
\alias{surf.dat.generate}
\title{Create SURF.DAT}
\usage{
surf.dat.generate(startyear = NOAA_start_year, endyear = NOAA_end_year,
  outputfile = "surf.dat")
}
\arguments{
\item{startyear}{the starting year for the SURF.DAT file.}

\item{endyear}{the ending year for the SURF.DAT file.}

\item{outputfile}{the specified filename for the SURF.DAT file. Defaults to "surf.dat".}
}
\description{
Generate a SURF.DAT input file for CALMET
}
\examples{
\dontrun{
# Generate "surf.dat" from previously defined period
surf.dat.generate()
}
}

